# server/openapi.yaml
openapi: "3.0.3"
info:
  title: "GIAM-SAT API v1.13.0"
  description: |
    ## GIAM-SAT Security Monitoring System

    REST API for the GIAM-SAT security monitoring platform.
    
    **Features:**
    - Agent monitoring & management
    - Event log collection & analysis  
    - File integrity monitoring (FIM)
    - Network traffic capture (Wireshark-level)
    - Vulnerability scanning (NVD + CVE)
    - YARA malware detection
    - SCA/CIS compliance assessment
    - Threat intelligence (OTX + VirusTotal + URLhaus + MISP)
    - AI-powered security analysis (DeepSeek/OpenAI/Gemini/Groq/xAI)
    - Telegram Bot notifications
    - Agentless monitoring (SNMP/SSH/Ping)
    
    **Authentication:** JWT Bearer token (except localhost)
    
    **Rate Limits:** 5 req/min for `/api/float_ai_chat`
    
    **Base URL:** `http://{host}:5000`
    
  version: "1.13.0"
  contact:
    name: "GIAM-SAT"
  license:
    name: "Proprietary"
servers:
  - url: "http://localhost:5000"
    description: "Local development server"
  - url: "http://{server_ip}:5000"
    description: "Production server"
    variables:
      server_ip:
        default: "YOUR_SERVER_IP"
        description: "GIAM-SAT server IP address"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token from /api/auth/login"
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
        machine_id:
          type: string
        event_id:
          type: string
          example: "4625"
        channel:
          type: string
          example: "Security"
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW, INFO]
        description:
          type: string
        timestamp:
          type: string
          format: date-time
    NetworkPacket:
      type: object
      properties:
        id:
          type: integer
        machine_id:
          type: string
        src_ip:
          type: string
          example: "192.168.1.100"
        dst_ip:
          type: string
          example: "8.8.8.8"
        src_port:
          type: integer
          example: 54321
        dst_port:
          type: integer
          example: 53
        protocol:
          type: string
          example: "UDP"
        protocol_app:
          type: string
          example: "DNS"
        dns_info:
          type: object
          properties:
            qname:
              type: string
            qtype_name:
              type: string
            rcode_name:
              type: string
        http_info:
          type: object
          properties:
            http_method:
              type: string
            http_uri:
              type: string
        raw_data:
          type: string
          description: "Full packet JSON with hex dump"
        timestamp:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /api/events:
    get:
      tags: [Events]
      summary: "Get event logs"
      parameters:
        - name: machine_id
          in: query
          schema:
            type: string
          description: "Filter by machine"
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: "Max results"
        - name: severity
          in: query
          schema:
            type: string
            enum: [CRITICAL, HIGH, MEDIUM, LOW, INFO]
        - name: event_id
          in: query
          schema:
            type: string
          description: "Filter by Windows Event ID"
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: "Events after this timestamp"
      responses:
        "200":
          description: "List of events"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
...
  /api/fim:
    get:
      tags: [FIM]
      summary: "Get File Integrity Monitoring events"
      parameters:
        - name: machine_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: action
          in: query
          schema:
            type: string
            enum: [FILE_CREATED, FILE_MODIFIED, FILE_DELETED]
      responses:
        "200":
          description: "List of FIM events"
...
  /api/network:
    get:
      tags: [Network]
      summary: "Get network traffic logs with application layer data"
      parameters:
        - name: machine_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: protocol
          in: query
          schema:
            type: string
            enum: [TCP, UDP, ICMP]
        - name: protocol_app
          in: query
          schema:
            type: string
            enum: [DNS, HTTP, TLS]
          description: "Filter by application layer protocol"
      responses:
        "200":
          description: "List of network packets with application layer info"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NetworkPacket"
...